+++ spec/jwt/jwa/hmac_spec.rb
@@ -17,26 +17,26 @@ RSpec.describe JWT::JWA::Hmac do
     context 'when nil hmac_secret is passed' do
       let(:hmac_secret) { nil }
 
-      it 'raises JWT::VerificationError' do
-        expect { subject }.to raise_error(JWT::VerificationError, 'HMAC key expected to be a String')
+      it 'raises JWT::EncodeError' do
+        expect { subject }.to raise_error(JWT::EncodeError, 'HMAC key expected to be a String')
       end
 
       it 'does not call OpenSSL::HMAC.digest' do
         expect(OpenSSL::HMAC).not_to receive(:digest)
-        expect { subject }.to raise_error(JWT::VerificationError)
+        expect { subject }.to raise_error(JWT::EncodeError)
       end
     end
 
     context 'when blank hmac_secret is passed' do
       let(:hmac_secret) { '' }
 
-      it 'raises JWT::VerificationError' do
-        expect { subject }.to raise_error(JWT::VerificationError, 'HMAC key cannot be empty')
+      it 'raises JWT::EncodeError' do
+        expect { subject }.to raise_error(JWT::EncodeError, 'HMAC key cannot be empty')
       end
 
       it 'does not call OpenSSL::HMAC.digest' do
         expect(OpenSSL::HMAC).not_to receive(:digest)
-        expect { subject }.to raise_error(JWT::VerificationError)
+        expect { subject }.to raise_error(JWT::EncodeError)
       end
     end
 
@@ -85,7 +85,7 @@ RSpec.describe JWT::JWA::Hmac do
         let(:hmac_secret) { 'short' }
 
         it 'raises error' do
-          expect { subject }.to raise_error(JWT::VerificationError, 'HMAC key must be at least 32 bytes for HS256 algorithm')
+          expect { subject }.to raise_error(JWT::EncodeError, 'HMAC key must be at least 32 bytes for HS256 algorithm')
         end
       end